feat: добавить валидацию. токен, авторизацию

+++ controllers/cards.js
@@ -1,6 +1,7 @@
 const {
   HTTP_STATUS_NOT_FOUND,
   HTTP_STATUS_CREATED,
+  HTTP_STATUS_FORBIDDEN,
 } = require('node:http2').constants;
 const createError = require('http-errors');
 const cardModel = require('../models/card');
@@ -15,8 +16,11 @@ const deleteCard = asyncHandler(async (req, res, next) => {
   const card = await cardModel
     .findById(req.params.cardId)
     .orFail(() => next(createError(HTTP_STATUS_NOT_FOUND, 'Карточка не найдена')));
+  if (card.owner._id !== req.user._id) {
+    return next(createError(HTTP_STATUS_FORBIDDEN, 'Нет прав на удаление карточки'));
+  }
   await cardModel.deleteOne(card);
-  res.send({});
+  return res.send({});
 });
 
 const createCard = asyncHandler(async (req, res) => {
